Improve parameter names and fix startup method arguments
We did not really support any method parameters that were not strings,
since we were getting the parameter type from the type of the string
split with `method:arg1:arg2:argN`, which was obviously going to be
an array of string types. So if the method did need type conversion,
we'd actually fail to locate it.

It was also hard to diagnose so added much more logging to the method
resolving and parameter conversions, as well as logging of the actual
parameter values used in the invocation.